+++ finance_modules/SpendingCategoryPieChart.jsx
