+++ finance_modules/SpendingMonthlyBarChart.jsx
